refactor: controller 네이밍 consoleController로 변경
