+++ src/main/java/org/prgms/kdtspringweek1/consoleController/CustomerController.java
@@ -1,7 +1,7 @@
-package org.prgms.kdtspringweek1.controller;
+package org.prgms.kdtspringweek1.consoleController;
 
 import org.prgms.kdtspringweek1.console.ConsoleOutput;
-import org.prgms.kdtspringweek1.controller.dto.SelectFunctionTypeDto;
+import org.prgms.kdtspringweek1.consoleController.dto.SelectFunctionTypeDto;
 import org.prgms.kdtspringweek1.customer.service.dto.FindCustomerResponseDto;
 import org.prgms.kdtspringweek1.customer.service.dto.UpdateCustomerRequestDto;
 import org.prgms.kdtspringweek1.customer.entity.Customer;
